Split gradle_plugin_test.dart
